+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-04-09T04:24:20.915Z",
+    "2026-04-09T12:48:37.295Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {